+++ CHANGES.rst
@@ -7,6 +7,8 @@
 
 - Drop support for Python 2.7, 3.5, 3.6.
 
+- Drop support for deprecated ``python setup.py test``.
+
 - Add support for Python 3.11.